+++ api/blueprints/__init__.py
@@ -1,3 +1,4 @@
+from .drbCitation import citation
 from .drbCollection import collection
 from .drbEdition import edition
 from .drbInfo import info